# swagger.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Product Management API'
  description: >
    # Welcome!

    This is a generalized Product Management API that can be integrated with an application for ease in product management.  
    
    It can be used for management of products that belong to, single category (like a pharmacy store) or multiple categories (like ebay).  

    ___
    
    ### Functions:

    * Basic CRUD functionalities of get, add, modify or remove products / categories. 
    

    * Products support extra functionalities like sort, filter and search. These can be helpful if API is integrated into an application.

    * Product search is dynamic and can display products which match even some part of query.

    * Categories are placed in a tree-like data-structure. If deleted, child categories of the root will move one level up in the tree.
    
    * For security to be added, API-KEY security is added to protect the API from unauthorized users.

    * This API follows the [RESTful architecture](https://restfulapi.net/)

  contact:
    name: Mridul Gupta
    email: mridul.gupta2021@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/mgvit2021/ProductManagementSystemAPI/1.0.0
  - description: Heroku test server for the API.
    url: https://pms-api-v1.herokuapp.com/
...
security:
  - ApiKeyAuth: []
paths:
...
  /products:
    get:
      summary: Get all the products available.
      tags:
        - Product Routes
      operationId: getAllProducts
      description: >
        **Get information about all the products from the store. Returns an object with status and products list.**
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Error while fetching products.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Adding a new product to the store.
      description: Creating a new product.
      operationId: createProduct
      tags:
        - Product Routes
      requestBody:
        description: Product information can be added through body as json-object or through form (if used) as well.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostProduct'
          # application/x-www-form-urlencoded:
          #   schema:
          #     $ref: '#/components/schemas/PostProduct'
      responses:
        '201':
          description: 'Successfully created a new employee object'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/MessageResponse'
                properties:
                  value:
                    $ref: '#/components/schemas/PostCategory'
                  links:
                     $ref: '#/components/schemas/LinkDescription'
                  
                  
        '400':
          description: An error while creating the product
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
           
        '404':
          description: Resource not found error
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/MessageResponse'
  /products/sort:
    get:
      summary: Sort products collection.
      description: >
        **Products can be sorted by base-price, discount and date-added. Sort order defines the order in which you want to display the products.**
      operationId: sortProducts
      tags:
        - Product Operations
      parameters: 
        - in: query
          name: sort_by
          description: Parameter by which you want the collection to be sorted. 
          schema:
            type: string
            enum:
              - basePrice
              - discount
              - created
          required: true
        - in: query
          name: sort_order
          description: Order in which you want the results to be displayed.
          schema:
            type: string
            enum: 
              - asc
              - desc
            description: >
              Sort order:
               * `asc` - Ascending, from A to Z
               * `desc` - Descending, from Z to A
          required: true
      responses:
        '200':
          description: All the products are returned in given sorted order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Operation unsuccesfull due to bad query formation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
  /products/filter:
    get:
      summary: Filters the products based on multiple query parameters.
      description: >
        **Fetches all the products filtered by query parameters(Requires atleast one query parameter). Returns an object with status and list of products.**
      operationId: filterProducts
      tags:
        - Product Operations
      parameters: 
        - in: query
          name: brand
          description: Product brand
          schema:
            type: string
        - in: query
          name: category_name
          description: Product category
          schema:
            type: string
        - in: query
          name: color
          description: Product color
          schema:
            type: string
        - in: query
          name: outOfStock
          description: Available products
          schema:
            type: boolean

      responses:
        '200':
          description: Returns filtered products.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Operation unsuccesfull due to bad query formation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
           
        '404':
          description: No products returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /products/search:
    get:
      summary: Search products by their name or by brand.
      description: >
        **Returns list of the products with the given name/brand, even if some match exists (regex match). Can be used with search-bars in applications.**
      operationId: searchProducts
      tags:
        - Product Operations
      parameters: 
        - in: query
          name: name
          description: Product name
          schema:
            type: string
        - in: query
          name: brand
          description: Product brand
          schema:
            type: string
      responses:
        '200':
          description: Returns list of matched products.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Bad query formation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"

        '404':
          description: No matching products found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /products/{prodId}:
    get:
      summary: Get a particular product.
      description: Returns product details of a product with given id.
      operationId: getProduct
      tags:
        - Product Routes
      parameters:
        - in: path
          name: prodId
          required: true
          description: The id of the product
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Product"
                properties: 
                  links:
                    $ref: "#/components/schemas/LinkDescription"
        '401':
           $ref: "#/components/responses/UnauthorizedError"

        '404':
          description: Product not found (Invalid id).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    put:
      summary: Update product details.
      description: Update product
      operationId: updateProduct
      tags:
        - Product Routes
      parameters:
        - in: path
          name: prodId
          required: true
          description: The id of the employee
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        description: Product information will be added through body as json-object.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostProduct'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PostProduct'
      responses:
        '200':
          description: Employee updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete a product.
      description: Delete a product.
      operationId: deleteProduct
      tags:
        - Product Routes
      parameters:
        - in: path
          name: prodId
          required: true
          description: The id of the employee
          schema:
            type: integer
            format: int64
            example: 23
      responses:
        '200':
          description: Product was deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Product could not be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: '#/components/responses/UnauthorizedError'
...
  /categories:
    get:
      summary: Get all the available categories.
      tags:
        - Category Routes
      description: Get all the categories available.
      operationId: getAllCategories
      responses:
        '200':
          description: Successfully retrieved all categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
              
        '401':
           $ref: "#/components/responses/UnauthorizedError"
          
    post:
      summary: Creates a new category.
      tags:
        - Category Routes
      description: Create a new category by passing details in body as json object.
      operationId: createCategory
      requestBody:
        description: Information about category to be added.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCategory'

      responses:
        '201':
          description: 'Successfully created a new category'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/MessageResponse'
                properties:
                  value:
                    $ref: '#/components/schemas/PostCategory'
                  links:
                     $ref: '#/components/schemas/LinkDescription'
        '400':
          description: Error while creating category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
           $ref: "#/components/responses/UnauthorizedError"
...
components:
  schemas:
...
    PostProduct:
      type: object
      description: 'Model containing product details.'
      properties:
        name:
          type: string
        brand:
          type: string
        basePrice:
          type: integer
          format: int64
        category_name:
          type: string
        details:
          type: object
        inStock:
          type: integer
          format: int64
          minimum: 0
      required:
        - name
        - brand
        - basePrice
        - category_name
        - inStock
    ProductResponse:
      type: object
      properties:
        status:
          type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...
    MessageResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        X-API-KEY:
          description: Header value
          schema:
            type: string
#No authentication in the test-API yet.
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key